backstage: remove unused JS from share.js and relocate functions responsible for generating links
