+++ components/o-share/src/tsx/share.tsx
@@ -1,5 +1,3 @@
-import {generateSocialUrl, generateDescriptiveLinkText} from '../js/share';
-
 type SocialNetworkType =
 	| 'twitter'
 	| 'facebook'
@@ -50,7 +48,7 @@ export function Share({
 							)}
 							rel="noopener">
 							<span className="o-share__text">
-								{generateDescriptiveLinkText({title}, socialNetwork)}
+								{generateDescriptiveLinkText(title, socialNetwork)}
 							</span>
 						</a>
 					</li>
@@ -58,4 +56,36 @@ export function Share({
 			</ul>
 		</div>
 	);
+}
+
+
+function generateDescriptiveLinkText (title: string, socialNetwork: SocialNetworkType) {
+	const descriptiveLinkText = {
+		twitter: `Share ${title} on Twitter (opens a new window)`,
+		facebook: `Share ${title} on Facebook (opens a new window)`,
+		linkedin: `Share ${title} on LinkedIn (opens a new window)`,
+		pinterest: `Share ${title} on Pinterest (opens a new window)`,
+		whatsapp: `Share ${title} on Whatsapp (opens a new window)`,
+		link: `Open link in new window`,
+		enterpriseSharing: `Share ${title} with your Enterprise Sharing tools (opens a new window)`,
+	};
+	return descriptiveLinkText[socialNetwork];
+}
+
+function generateSocialUrl(config: Record<string, string>, socialNetwork: SocialNetworkType) {
+	const url = encodeURIComponent(config.url);
+	const title = encodeURIComponent(config.title);
+	const titleExtra = encodeURIComponent(config.titleExtra);
+	const summary = encodeURIComponent(config.summary);
+	const relatedTwitterAccounts = encodeURIComponent(config.relatedTwitterAccounts);
+	const socialUrls = {
+		twitter: `https://twitter.com/intent/tweet?url=${url}&text=${title}&related=${relatedTwitterAccounts}&via=FT`,
+		facebook: `http://www.facebook.com/sharer.php?u=${url}`,
+		linkedin: `http://www.linkedin.com/shareArticle?mini=true&url=${url}&title=${title}+%7C+${titleExtra}&summary=${summary}&source=Financial+Times`,
+		pinterest: `http://www.pinterest.com/pin/create/button/?url=${url}&description=${title}`,
+		whatsapp: `whatsapp://send?text=${title}%20(${titleExtra})%20-%20${url}`,
+		link: url,
+		enterpriseSharing: url,
+	};
+	return socialUrls[socialNetwork];
 }
